+++ src/network.ts
@@ -3,5 +3,5 @@ import { PublicKey } from "@solana/web3.js";
 export enum Network {
   LOCALNET = "localnet",
   DEVNET = "devnet",
-  MAINNET = "mainnet_beta",
+  MAINNET = "mainnet",
 }